Add awk script for parsing dataSet.dat
This was initially just a temporary file while I worked on optimizing
your use of `awk`, but it turned into something more. I hope this is
both educational and otherwise useful.

See the comments in the file (especially its header) for more
information.

I only ask that you please keep its header, or at least the line
stating the author and the E-Mail address.

If you have any questions, I'm happy to help.