Fix up, optimize, make robust, tidy, etc
A _lot_ has been done, but it's certainly not completely yet.

This is a basically a massive overhaul, as the script was an absolute
mess beforehand, if I'm brutally honest.

I'm fixing up the broken indenting (using tabs; can change this if
desired) and ensuring they are consistent, as any language should be,
in my view.

A _lot_ of awk optimizations have been made, including writing up a new
awk script which is available on the master branch, as `parsedb.awk`
and will be called upon within the `tool.sh` script.

Argument processing is far cleaner now, using a standard `while`, -
`case`, `shift` situation. Also reworking the logic to better fit, be
more concise, and be easier to edit.

Since it's not the shortest script ever, I've taken to a more condensed
writing style, to avoid an excessive line count.

That being said, a lot of the comments amended to the lines were moved
to a more traditional place of above the related code; this allows for
easier reading, at the expense of more lines through which to scroll.

These are probably just some of the changes.